Update card border colors
