+++ src/app/GameBoard/page.tsx
@@ -62,7 +62,7 @@ const GameBoard = () => {
             mr: sidebarOpen ? `${drawerWidth}px` : '0',
             height: '100vh',
             position: 'relative',
-            backgroundImage: `url(${s3ImageURL('game/board-background-1.png')})`,
+            backgroundImage: `url(${s3ImageURL('ui/board-background-1.webp')})`,
             backgroundSize: 'cover',
             backgroundPosition: 'center',
         },
