+++ src/app/_components/Gameboard/Board/Board.tsx
@@ -3,7 +3,7 @@ import Grid from '@mui/material/Grid2';
 import UnitsBoard from '../_subcomponents/UnitsBoard';
 import { IBoardProps } from '@/app/_components/Gameboard/GameboardTypes';
 import { useGame } from '@/app/_contexts/Game.context';
-import LeaderBaseCard from '@/app/_components/_sharedcomponents/Cards/LeaderBaseCard/LeaderBaseCard';
+import LeaderBaseCard from '@/app/_components/_sharedcomponents/Cards/LeaderBaseCard';
 import { Box } from '@mui/material';
 
 const Board: React.FC<IBoardProps> = ({
