+++ src/app/_components/Lobby/Players/Players.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { Card, Box, Typography } from '@mui/material';
 import Grid from '@mui/material/Grid2';
 import { ILobbyUserProps, IPlayersProps } from '../LobbyTypes';
-import LeaderBaseCard from '@/app/_components/_sharedcomponents/Cards/LeaderBaseCard/LeaderBaseCard';
+import LeaderBaseCard from '@/app/_components/_sharedcomponents/Cards/LeaderBaseCard';
 import { useGame } from '@/app/_contexts/Game.context';
 
 const Players: React.FC<IPlayersProps> = ({ isLobbyView }) => {
