+++ src/app/lobby/page.tsx
@@ -26,7 +26,7 @@ const Lobby = () => {
         containerStyle: {
             height: '100vh',
             overflow: 'hidden',
-            backgroundImage: `url(${s3ImageURL('game/board-background-1.png')})`,
+            backgroundImage: `url(${s3ImageURL('ui/board-background-1.webp')})`,
             backgroundSize: 'cover',
             backgroundPosition: 'center',
         },
